+++ package.json
@@ -44,24 +44,24 @@
   },
   "dependencies": {
     "class-transformer": "^0.4.0",
-    "typeorm": "^0.2.31"
+    "typeorm": "^0.2.34"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.11",
-    "@types/chai-as-promised": "^7.1.2",
-    "@types/mocha": "^8.2.1",
-    "@types/node": "^14.14.35",
-    "chai": "^4.2.0",
+    "@types/chai": "^4.2.21",
+    "@types/chai-as-promised": "^7.1.4",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^16.4.1",
+    "chai": "^4.3.4",
     "chai-as-promised": "^7.1.1",
-    "mocha": "^8.3.2",
-    "mongodb": "^3.5.6",
-    "mssql": "^6.2.0",
+    "mocha": "^9.0.2",
+    "mongodb": "^4.0.1",
+    "mssql": "^7.1.3",
     "mysql": "^2.18.1",
-    "pg": "^8.0.2",
+    "pg": "^8.6.0",
     "rimraf": "^3.0.2",
     "sqlite3": "^5.0.2",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.1.0",
     "tslint": "^6.1.1",
-    "typescript": "^4.2.3"
+    "typescript": "^4.3.5"
   }
 }
